requirements.txt
Flask>=3.0
python-docx>=1.1
openpyxl>=3.1
lxml>=5.0
